+++ src/drivers/exec/exec.c
@@ -551,7 +551,7 @@ ssize_t exec_handler(context_t *ctx, event_t event, driver_data_t *event_data )
 		case EVENT_READ:
 			{
 				size_t bytes;
-				event_bytes( fd->fd, &bytes );
+				event_bytes( (int)fd->fd, &bytes );
 
 				if( bytes ) {
 					char read_buffer[MAX_READ_BUFFER];
@@ -559,7 +559,7 @@ ssize_t exec_handler(context_t *ctx, event_t event, driver_data_t *event_data )
 					if( bytes >= MAX_READ_BUFFER )
 						bytes = MAX_READ_BUFFER-1;
 
-					ssize_t result = event_read( fd->fd, read_buffer, bytes);
+					ssize_t result = event_read( (int)fd->fd, read_buffer, bytes);
 
 					if( result >= 0 ) {
 						read_buffer[result] = 0;
@@ -572,7 +572,7 @@ ssize_t exec_handler(context_t *ctx, event_t event, driver_data_t *event_data )
 							x_printf(ctx,"WARNING: Failed to send all data to %s\n",ctx->owner->name);
 
 					} else
-						x_printf(ctx," * WARNING: read return unexpected result %d\n",result);
+						x_printf(ctx," * WARNING: read return unexpected result %d\n",(int)result);
 				} else {
 					x_printf(ctx,"EOF on input fd %ld\n",fd->fd);
